+++ Sources/TuistGraph/Models/BuildAction.swift
@@ -7,15 +7,18 @@ public struct BuildAction: Equatable {
     public var targets: [TargetReference]
     public var preActions: [ExecutionAction]
     public var postActions: [ExecutionAction]
+    public var runPostActionsOnFailure: Bool
 
     // MARK: - Init
 
     public init(targets: [TargetReference] = [],
                 preActions: [ExecutionAction] = [],
-                postActions: [ExecutionAction] = [])
+                postActions: [ExecutionAction] = [],
+                runPostActionsOnFailure: Bool = false)
     {
         self.targets = targets
         self.preActions = preActions
         self.postActions = postActions
+        self.runPostActionsOnFailure = runPostActionsOnFailure
     }
 }